update to angular 13
library build with ng-packagr
