+++ .gitignore
@@ -1,52 +1,46 @@
-# Logs
-logs
-*.log
-npm-debug.log*
+# See http://help.github.com/ignore-files/ for more about ignoring files.
 
-# Runtime data
-pids
-*.pid
-*.seed
+# compiled output
+/dist
+/tmp
+/out-tsc
+# Only exists if Bazel was run
+/bazel-out
 
-.vscode
-dist
-*.metadata.json
-*.ngsummary.json
-*.ngfactory.ts
+# dependencies
+/node_modules
 
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
+# profiling files
+chrome-profiler-events*.json
 
-# Coverage directory used by tools like istanbul
-coverage
+# IDEs and editors
+/.idea
+.project
+.classpath
+.c9/
+*.launch
+.settings/
+*.sublime-workspace
 
-# nyc test coverage
-.nyc_output
+# IDE - VSCode
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+.history/*
 
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
+# misc
+/.angular/cache
+/.sass-cache
+/connect.lock
+/coverage
+/libpeerconnection.log
+npm-debug.log
+yarn-error.log
+testem.log
+/typings
 
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
-# Dependency directories
-node_modules
-jspm_packages
-
-# Optional npm cache directory
-.npm
-
-# Optional REPL history
-.node_repl_history
-
-# Generated #
-!karma.conf.js
-!karma-test-shim.js
-*.js
-!rollup.config.js
-*.map
-*.d.ts
-typings
+# System Files
+.DS_Store
+Thumbs.db
