--- ng2-img-max.ts
@@ -1,5 +1,0 @@
-export { Ng2ImgMaxService } from './src/ng2-img-max.service';
-export { Ng2ImgMaxModule } from './src/ng2-img-max.module';
-export { ImgMaxSizeService } from './src/img-max-size.service';
-export { ImgMaxPXSizeService } from './src/img-maxpx-size.service';
-export { ImgExifService } from './src/img-exif.service';
