+++ package.json
@@ -1,63 +1,50 @@
 {
-    "name": "ng2-img-max",
-    "version": "2.2.4",
-    "description": "Angular module to resize images down to a certain width and height or to reduce the quality to fit a certain maximal filesize - all in the browser.",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/bergben/ng2-img-max.git"
-    },
-    "license": "MIT",
-    "main": "./dist/bundles/ng2-img-max.umd.js",
-    "module": "./dist/ng2-img-max.js",
-    "typings": "./dist/ng2-img-max.d.ts",
-    "scripts": {
-        "transpile": "ngc",
-        "package": "rollup -c",
-        "minify": "uglifyjs dist/bundles/ng2-img-max.umd.js --screw-ie8 --compress --mangle --comments --output dist/bundles/ng2-img-max.umd.min.js",
-        "build": "npm run transpile && npm run package && npm run minify"
-    },
-    "author": {
-        "name": "Benedikt Berger",
-        "email": "benedikt.berger@live.de"
-    },
-    "keywords": [
-        "angular",
-        "angular2",
-        "img-max",
-        "maximal",
-        "size",
-        "limit",
-        "resize"
-    ],
-    "dependencies": {
-        "exifr": "^4.3.4",
-        "ng2-pica": "^1.3.1"
-    },
-    "peerDependencies": {
-        "@angular/core": "^6.0.4",
-        "core-js": ">=2.4.1",
-        "rxjs": "^6.2.0"
-    },
-    "devDependencies": {
-        "@angular/common": "^6.0.9",
-        "@angular/compiler": "^6.0.9",
-        "@angular/compiler-cli": "^6.0.9",
-        "@angular/core": "^6.0.9",
-        "@angular/platform-browser": "^6.0.9",
-        "@angular/platform-browser-dynamic": "^6.0.9",
-        "@angular/router": "^6.0.9",
-        "@types/node": ">=7.0.5",
-        "core-js": "^2.4.1",
-        "reflect-metadata": "^0.1.3",
-        "rollup": "^0.41.4",
-        "rxjs": "^6.2.0",
-        "tslint": "^4.4.2",
-        "typescript": "2.7.2",
-        "uglify-js": "^2.7.5",
-        "zone.js": "^0.8.26"
-    },
-    "bugs": {
-        "url": "https://github.com/bergben/ng2-img-max/issues"
-    },
-    "homepage": "https://github.com/bergben/ng2-img-max#readme"
+  "name": "ng2-img-max",
+  "version": "2.2.5",
+  "description": "Angular module to resize images down to a certain width and height or to reduce the quality to fit a certain maximal filesize - all in the browser.",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/bergben/ng2-img-max.git"
+  },
+  "license": "MIT",
+  "scripts": {
+    "build": "ng build"
+  },
+  "author": {
+    "name": "Benedikt Berger",
+    "email": "benedikt.berger@live.de"
+  },
+  "keywords": [
+    "angular",
+    "angular2",
+    "img-max",
+    "maximal",
+    "size",
+    "limit",
+    "resize"
+  ],
+  "peerDependencies": {
+    "@angular/common": "^13.1.2",
+    "@angular/compiler": "~13.1.2",
+    "@angular/core": "^13.1.2",
+    "rxjs": "~7.5.2"
+  },
+  "dependencies": {
+    "exifr": "^4.3.4",
+    "pica": "^9.0.1",
+    "tslib": "^2.3.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.2",
+    "@types/node": "^12.11.1",
+    "ng-packagr": "^13.1.1",
+    "typescript": "~4.5.2"
+  },
+  "bugs": {
+    "url": "https://github.com/bergben/ng2-img-max/issues"
+  },
+  "homepage": "https://github.com/bergben/ng2-img-max#readme",
+  "private": false
 }
