+++ src/public-api.ts
@@ -1,0 +1,5 @@
+export * from './lib/img-exif.service';
+export * from './lib/img-max-size.service';
+export * from './lib/img-maxpx-size.service';
+export * from './lib/ng2-img-max.module';
+export * from './lib/ng2-img-max.service';
